separate txes and queries by module
